[FE] Feat: 맵 컴포넌트 경로 표시하기 (#79)
* Fix: 경로 계산 식 수정 (#72)

- d / 1000 에서 d * 1000 으로 수정

* Fix: 파일 import 경로 수정 (#72)

- Modal.module.scss, Test.module.scss 불러오는 과정에서 대소문자 틀림
- 수정 완료

* Fix: currentLocation -> location으로 변경 (#72)

- watchPosition 내부에서 currentLocation을 모르는 이슈 발생
- 갱신된 currentLocation 값 대신 location을 사용하는 방식으로 변경

* Refactor: 구글맵 옵션과 거리 계산 함수 분리 작업 (#72)

- google map style 옵션, getDistance 함수 utils 폴더로 분리
