+++ client/src/components/common/Map.tsx
@@ -1,6 +1,8 @@
 import { GoogleMap, LoadScript, Marker, Polyline } from '@react-google-maps/api'
 import { useState, useEffect } from 'react'
 import styles from './Map.module.scss'
+import getDistance from '../../utils/calculate'
+import { googleMapOptions, polylineOptions } from '../../utils/options'
 
 // 좌표 테스트 ing
 
@@ -19,11 +21,7 @@ function Map() {
   const [isLoadMap, setIsLoadMap] = useState(false)
   const [distance, setDistance] = useState(0)
 
-  console.log(locations)
-  console.log(currentLocation)
-
   useEffect(() => {
-    setIsLoadMap(true)
     setApiKey(import.meta.env.VITE_GOOGLE_MAPS_API_KEY || '')
     navigator.geolocation.getCurrentPosition(
       position => {
@@ -40,49 +38,39 @@ function Map() {
       },
       { timeout: 5000 }
     )
-  }, [])
-
-  function getDistance(lat1: number, lng1: number, lat2: number, lng2: number): number {
-    function deg2rad(deg: number): number {
-      return deg * (Math.PI / 180)
+    setIsLoadMap(true)
+    return () => {
+      setIsLoadMap(false)
     }
-    const R = 6371 // Radius of the earth in km
-    const dLat = deg2rad(lat2 - lat1) // deg2rad below
-    const dLon = deg2rad(lng2 - lng1)
-    const a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) * Math.sin(dLon / 2) * Math.sin(dLon / 2)
-    const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a))
-    const d = (R * c) / 1000 // Distance in m
-    return d
-  }
+  }, [])
 
   function watchLocation() {
     let prevLocation: LatLng | undefined // 이전 위치
 
     if (navigator.geolocation) {
-      console.log('geolocation 실행')
       // 기기의 현재 위치를 탐색하는 브라우저 api 사용
       navigator.geolocation.watchPosition(
         position => {
+          console.log('watchPosition 실행')
           // 좌표를 담아둘 객체 정의
           const location = {
             lat: position.coords.latitude,
             lng: position.coords.longitude,
           }
           // 현재 위치 갱신
-          setCurrentLocation(location)
+          setCurrentLocation(() => location) // state
           // 직전 좌표가 없으면 현재 watchPosition으로 받은 location을 locations 배열에 추가
           if (!prevLocation) {
             setLocations([location])
             prevLocation = location
           }
           // 직전 좌표가 있으면 locations 배열에 현재 위치 추가
-          if (currentLocation && prevLocation) {
+          // currentLocation에서 location으로 변경
+          if (location && prevLocation) {
             // 직전 좌표가 존재하면 현재 좌표와 직전 좌표의 거리를 계산
             const distanceFromLastPosition = getDistance(
-              currentLocation.lat,
-              currentLocation.lng,
+              location.lat,
+              location.lng,
               prevLocation.lat,
               prevLocation.lng
             )
@@ -92,14 +80,15 @@ function Map() {
             if (distanceFromLastPosition >= 5) {
               setLocations(prevLocations => [...prevLocations, location])
               prevLocation = location
+              setDistance(0)
             }
           }
         },
         err => {
-          console.log(err)
+          console.log('watchPosition Error :', err)
         },
         {
-          enableHighAccuracy: false,
+          enableHighAccuracy: true,
           maximumAge: 0,
           timeout: 5000, // 5초마다 위치 정보 갱신 시도
         }
@@ -107,7 +96,6 @@ function Map() {
     } else {
       console.log('Geolocation is not supported')
     }
-    // vercel에서 설정한 환경 변수를 가져옴
   }
 
   return (
@@ -118,62 +106,10 @@ function Map() {
             mapContainerStyle={containerStyle}
             zoom={14}
             center={currentLocation}
-            options={{
-              styles: [
-                {
-                  // 색상
-                  featureType: 'all',
-                  stylers: [
-                    {
-                      saturation: -100,
-                    },
-                  ],
-                },
-                {
-                  // 물 색상
-                  featureType: 'water',
-                  stylers: [
-                    {
-                      color: '#7dcdcd',
-                    },
-                  ],
-                },
-                {
-                  // 건물 이름 가리기
-                  featureType: 'all',
-                  elementType: 'labels',
-                  stylers: [{ visibility: 'off' }],
-                },
-
-                {
-                  // 지도 단순화하기
-                  featureType: 'road',
-                  elementType: 'geometry',
-                  stylers: [
-                    { visibility: 'simplified' },
-                    { hue: '#000000' },
-                    { saturation: -50 },
-                    { lightness: -15 },
-                    { weight: 1.5 },
-                  ],
-                },
-              ],
-              mapTypeControl: false, // 지도 위성 끄기
-              streetViewControl: false, // 사람 모양 끄기
-              gestureHandling: 'greedy', // 한 손 가락으로 지도 핸들링
-            }}
+            options={googleMapOptions}
           >
             {currentLocation && <Marker position={currentLocation} />}
-            {locations.length > 1 && (
-              <Polyline
-                path={locations}
-                options={{
-                  strokeColor: '#55de50',
-                  strokeOpacity: 1,
-                  strokeWeight: 6,
-                }}
-              />
-            )}
+            {locations.length > 1 && <Polyline path={locations} options={polylineOptions} />}
           </GoogleMap>
         </LoadScript>
       ) : null}
@@ -192,7 +128,7 @@ function Map() {
       </div>
       <div className={styles.distanceBox}>
         <div>직전 좌표와의 거리: </div>
-        <div>{distance.toFixed(2)}</div>
+        <div>{distance.toFixed(8)}</div>
         <div>m</div>
       </div>
       <div className={styles.locationInfoBox}>
